[cal]BM-17865 Fix: npe protection

+++ parent/videoconferencing/net.bluemind.videoconferencing.saas.service/src/net/bluemind/videoconferencing/saas/service/internal/VideoConferencingSaasService.java
@@ -310,7 +310,9 @@ public class VideoConferencingSaasService implements IVideoConferencingSaas, IIn
 	public void delete(String roomName) {
 		RBACManager.forContext(context).check("hasFullVideoconferencing", "hasSimpleVideoconferencing");
 		ItemValue<BlueMindVideoRoom> itemRoom = storeService.byIdentifier(roomName);
-		storeService.delete(itemRoom.uid);
+		if (itemRoom != null) {
+			storeService.delete(itemRoom.uid);
+		}
 	}
 
 }
